+++ tests/checks.py
@@ -14,7 +14,7 @@ class checks:
                     "find . -name '*.py' | xargs black -l 80",
                 ),
                 (
-                    dict(folder=["foo", "bar"]),
+                    dict(folders=["foo", "bar"]),
                     "find foo bar -name '*.py' | xargs black -l 79",
                 ),
                 (
@@ -30,7 +30,7 @@ class checks:
                         diff=True,
                         check=True,
                         line_length=80,
-                        folder=["foo", "bar"],
+                        folders=["foo", "bar"],
                     ),
                     "find foo bar -name '*.py' | xargs black -l 80 --check --diff",  # noqa
                 ),
@@ -48,14 +48,14 @@ class checks:
             blacken(ctx, **kwargs)
             ctx.run.assert_called_once_with(command, pty=True)
 
-        def folder_configurable(self, ctx):
+        def folders_configurable(self, ctx):
             # Just config -> works fine
             ctx.blacken = dict(folders=["elsewhere"])
             blacken(ctx)
             assert "elsewhere" in ctx.run.call_args[0][0]
 
-        def folder_config_loses_to_runtime(self, ctx):
+        def folders_config_loses_to_runtime(self, ctx):
             # Config + CLI opt -> CLI opt wins
             ctx.blacken = dict(folders=["nowhere"])
-            blacken(ctx, folder=["nowhere"])
+            blacken(ctx, folders=["nowhere"])
             assert "nowhere" in ctx.run.call_args[0][0]